backend/requirements.txt
fastapi
uvicorn
aiomysql
pymysql
passlib[bcrypt]
python-dotenv
httpx
pytest
pytest-asyncio
pytest-cov
slowapi
apscheduler
pandas
requests
tweepy
slack_sdk
twilio
gspread
google-auth
pyairtable
langchain
langchain_groq
python-jose
email-validator 